% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{format_dates}
\alias{format_dates}
\title{Takes an dataframe with a column containing ANDES database datetime strings and adds new columns for day, month, and year}
\usage{
format_dates(df, reference_column = "HEUR_DEB")
}
\arguments{
\item{df}{A dataframe containing a datetime string representing.}

\item{reference_column}{The name of the column that contains the ANDES UTC datetime string in "YYYY-MM-DD HH:MM:SS".}
}
\value{
A dataframe identical to the input dataframe with additional columns for day (JOUR), month (MOIS), and year (ANNEE).
}
\description{
Takes an dataframe with a column containing ANDES database datetime strings and adds new columns for day, month, and year
}
