% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsm.R
\name{get_fishing_sets_bsm}
\alias{get_fishing_sets_bsm}
\title{Gets fishing set activity}
\usage{
get_fishing_sets_bsm(andes_db_connection)
}
\arguments{
\item{andes_db_connection}{a connection object to the ANDES database.}
}
\value{
A dataframe containing fishing set data.
}
\description{
This function executes a SQL query to retrieve fishing set data from the ANDES database.
The current ANDES active mission will determine for which mission the sets are returned.
}
\seealso{
[get_fishing_sets_bsm_db()] for the raw database results
}
