% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsm.R
\name{get_specimen_bsm_db}
\alias{get_specimen_bsm_db}
\title{Gets specimen data (raw database results)}
\usage{
get_specimen_bsm_db(andes_db_connection)
}
\arguments{
\item{andes_db_connection}{a connection object to the ANDES database. Please}
}
\value{
A dataframe containing fishing set data.
}
\description{
This function executes a SQL query to retrieve specimen data from the ANDES database.
The current ANDES active mission will determine for which mission the specimens are returned.
}
\details{
This function is intended for internal use and returns raw results from the database.
It is not meant for direct use in analysis or reporting. Users should use `get_specimen_bsm`
}
\seealso{
[get_specimen_bsm()] for the formatted results
}
