% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{parse_andes_datetime}
\alias{parse_andes_datetime}
\title{Takes an ANDES database datetime string and converts it to a POSIXlt object}
\usage{
parse_andes_datetime(andes_time_str)
}
\arguments{
\item{andes_time_str}{A string representing a UTC date and time in the format "YYYY-MM-DD HH:MM:SS".}
}
\value{
A POSIXlt object representing the date and time in UTC.
}
\description{
Takes an ANDES database datetime string and converts it to a POSIXlt object
}
